requirements.txt
elasticsearch==6.3.1
elasticsearch_dsl==6.3.1
Django==2.1.9
django-widget-tweaks==1.4.3
django-queryset-csv==1.0.1
django-autofixture==0.12.1
django-extensions==2.1.5
wordcloud==1.5.0
nltk==3.3
pandas==0.24.2
matplotlib==3.0.3
whitenoise==4.1.2
django-plotly-dash==0.9.12
dpd-components==0.1.0
dpd-static-support==0.0.4
dash==0.43.0
dash-renderer==0.24.0
dash-core-components==0.48.0
dash-html-components==0.16.0
camelot-py==0.8.2
pypdf2==1.26
opencv-python==4.5.3.56
